requirements.txt
torch
Pillow
torchvision 
torchaudio 
scikit-learn